+++ client/CHANGELOG.md
@@ -28,7 +28,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 * Completion for Enumeration
 * Insider Mode for review procedure
-* [Insider] Goto Module -> Credit & thanks to [anacierdem](https://github.com/anacierdem) for letting me use his code on of his [Require Module Support Provider](https://marketplace.visualstudio.com/items?itemName=lici.require-js)
+* [Insider] Goto Module -> Credit & thanks to [anacierdem](https://github.com/anacierdem) for letting me use the code of his [Require Module Support Provider](https://marketplace.visualstudio.com/items?itemName=lici.require-js)
 
 ### Changed
 
@@ -39,6 +39,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 * Fixed element finding in XML Language Server
 * Fixed wrong types for elements in XML Language Server
 * Fixed attribute showing up when in attribute value section (part in quotes)
+* Fixed annoying manifest location found message, if manifest is set in workspace settings
 
 ## [0.3.1] - 2017-04-04